requirements.txt
Flask
Gunicorn
Schedule
Requests
Time
python-dotenv
logging
